Improvements for correctly filling in glyphs
Bug fix in BNode for MaxRight referencing wrong variable

TriangleFill to allow close triangle points to pass the inside-triangle
test. This is especially necessary when processing bridged geometry.

Cubic roots are refined on line-curve intersection. Without it,
degenerate situations can be drastically off and miss valid
intersections.

Boolean difference has options to do nothing if no collision occured -
which is neccessary to deffer the situation to the font bridging code.