Modified booleans and text for a more stable text implementation
Added a few more parameters for boolean functions including parameters to get an output node guaranteed to be on the left after the process in case the handle to the island is clipped.
Booleans for bridging text also have negatives only applied to one positive - the closest largest shape found around the negative.